add 0 and 1
