+++ app0.py
@@ -1,0 +1,38 @@
+from flask import Flask, jsonify
+from flask_restful import Resource, Api
+from flask_swagger_ui import get_swaggerui_blueprint
+import json
+
+app = Flask(__name__)
+api = Api(app)
+
+# Define a sample resource
+class HelloWorld(Resource):
+    def get(self):
+        return jsonify({'message': 'Hello, World!'})
+
+# Add the resource to the API
+api.add_resource(HelloWorld, '/hello')
+
+
+
+
+# Configure Swagger UI
+SWAGGER_URL = '/swagger'
+API_URL = 'http://127.0.0.1:5000/swagger0.json'
+swaggerui_blueprint = get_swaggerui_blueprint(
+    SWAGGER_URL,
+    API_URL,
+    config={
+        'app_name': "Sample API"
+    }
+)
+app.register_blueprint(swaggerui_blueprint, url_prefix=SWAGGER_URL)
+
+@app.route('/swagger0.json')
+def swagger():
+    with open('swagger0.json', 'r') as f:
+        return jsonify(json.load(f))
+
+if __name__ == '__main__':
+    app.run(debug=True)
